feat: run code locally
