+++ src/webCrawler/crawler.py
@@ -1,60 +1,29 @@
-#código funcionará no google colab
-
 import requests #importa biblioteca de requests get,post
 from bs4 import BeautifulSoup #importa bs4 para extrair os dados html e xml
 import pandas as pd #importa pd para manipular, organizar e exportar dados coletados em tabelas e salvar como csv
 from selenium import webdriver #simulador de navegador
 from selenium.webdriver.chrome.service import Service #faz com que o selenium simule um navegador
-from selenium.webdriver.chrome.options import Options #//
-import time #biblioteca de tempo para trabalhar com sleep(time.sleep(xxx)) e outras coisas
-from google.colab import files
+from selenium.webdriver.chrome.options import Options 
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.common.by import By
-from selenium.common.exceptions import TimeoutException
+import time #biblioteca de tempo para trabalhar com sleep(time.sleep(xxx)) e outras coisas
 import random #importa modulo de random
-import google.generativeai as genai #SDK da Gemini
-import ipywidgets as widgets #importa biblioteca para widgets
-from IPython.display import display #importa biblioteca para widgets
-from google.colab import userdata #importa o módulo que manipula variáveis de enviroment/secrets do colab
-#
+
 #credenciais do LambdaTest
 username = 'eduardo.evaristo.am'  #username
 access_key = 'VZGoC8JtKpO3xrvZTcIOTGaV3iFOP1YD2MnLguzKNGuT8N8QLv'  #access key
-#
-#configurações do ChromeOptions
-chrome_options = Options()
-chrome_options.add_argument("--start-maximized")  #inicia o navegador
-#
-#configurações do navegador
-capabilities = {
-    'LT:Options': {
-        'build': 'Selenium Python Sample',  #nome da sua build
-        'name': 'Sample Test',  #nome do seu teste
-        'platform': 'Windows 11',  #plataforma do SO
-        'browserName': 'Chrome',  #nome do navegador
-        'version': 'latest'  #versão do navegador
-    }
-}
 
-chrome_options.set_capability('LT:Options', capabilities['LT:Options'])
-#
-#conectando ao LambdaTest
-driver = webdriver.Remote(
-    command_executor=f'https://{username}:{access_key}@hub.lambdatest.com/wd/hub',
-    options=chrome_options
-)
-#
-# Configurações do ChromeOptions - Utilizar este
+
+# Configurações do ChromeOptions
 def config_driver():
   chrome_options = Options()
   chrome_options.add_argument("--start-maximized")  # Inicia o navegador
 
-  # Lista de user-agent strings
+  # Lista de user-agent strings para evitar cair no cloudflare
   user_agents = [
       "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.45 Safari/537.36",
       "Mozilla/5.0 (Windows NT 10.0; Win64; x64) Gecko/20100101 Firefox/94.0",
-      # Adicione mais user-agent strings conforme necessário
   ]
 
   # Definindo um user-agent aleatório
@@ -79,14 +48,14 @@ def config_driver():
       options=chrome_options
   )
   return driver
-#
-# Três páginas da zapimovieis
+
+# Três páginas da zapimoveis
 urls = {
     "zapimoveis1": "https://www.zapimoveis.com.br/aluguel/imoveis/es+vitoria/?__ab=sup-hl-pl:newC,exp-aa-test:control,super-high:new,off-no-hl:new,pos-zap:new,new-rec:b,ltroffline:control&transacao=aluguel&onde=,Esp%C3%ADrito%20Santo,Vit%C3%B3ria,,,,,city,BR%3EEspirito%20Santo%3ENULL%3EVitoria,-20.319664,-40.338475,&pagina=1",
     "zapimoveis2": "https://www.zapimoveis.com.br/aluguel/imoveis/es+vitoria/?__ab=sup-hl-pl:newC,exp-aa-test:control,super-high:new,off-no-hl:new,pos-zap:new,new-rec:b,ltroffline:control&transacao=aluguel&onde=,Esp%C3%ADrito%20Santo,Vit%C3%B3ria,,,,,city,BR%3EEspirito%20Santo%3ENULL%3EVitoria,-20.319664,-40.338475,&pagina=2",
     "zapimoveis3": "https://www.zapimoveis.com.br/aluguel/imoveis/es+vitoria/?__ab=sup-hl-pl:newC,exp-aa-test:control,super-high:new,off-no-hl:new,pos-zap:new,new-rec:b,ltroffline:control&transacao=aluguel&onde=,Esp%C3%ADrito%20Santo,Vit%C3%B3ria,,,,,city,BR%3EEspirito%20Santo%3ENULL%3EVitoria,-20.319664,-40.338475,&pagina=3"
 }
-#
+
 #função para buscar informações de um imóvel
 def get_imovel_data(url, driver):
     driver.get(url)
@@ -165,21 +134,20 @@ def get_imovel_data(url, driver):
     for listing in listings:
         print(listing)
     return listings
-#
+
 #função para coletar dados de todas as cidades e sites
 def coletar_dados_de_aluguel():
     dados_coletados = []
-    driver = config_driver()  # Inicie o driver uma vez
 
     for site, url in urls.items():
+        driver = config_driver()
         print(f"Coletando dados de {site}...")
         dados_imoveis = get_imovel_data(url, driver)
         dados_coletados.extend(dados_imoveis)
 
-    driver.quit()  # Feche o driver após a coleta
     return dados_coletados
 
-#
+
 #armazenar dados em um DataFrame e exportar
 dados_de_aluguel = coletar_dados_de_aluguel()
 df = pd.DataFrame(dados_de_aluguel)
@@ -187,45 +155,6 @@ df.to_csv('imoveis_aluguel.csv', index=True, encoding='utf-8')
 
 print("Coleta concluída e dados salvos em imoveis_aluguel.csv")
 print(df)
-files.download('imoveis_aluguel.csv') #descomentar caso queira baixar o csv
-#
-#Front-end
-api_key = userdata.get('API_KEY_GEMAI')
 
-genai.configure(api_key=api_key)
-model = genai.GenerativeModel("gemini-1.5-flash")
 
-# Criação de um campo de texto para a pergunta do usuário
-pergunta_texto = widgets.Text(
-    description='Pergunta:',
-    placeholder='Digite sua pergunta aqui...'
-)
 
-# Criação de um botão para enviar a pergunta
-botao_enviar = widgets.Button(description='Enviar')
-
-# Saída para mostrar a resposta
-saida_resposta = widgets.Output()
-
-#Lê arquivo csv
-with open('imoveis_aluguel.csv', 'r') as file:
-    dados = file.read()
-
-# Função para gerar a resposta quando o botão é clicado
-def on_botao_enviar_clicked(b):
-    with saida_resposta:
-        saida_resposta.clear_output()  # Limpa a saída anterior
-        pergunta = pergunta_texto.value
-        # Chamando o modelo com a pergunta do usuário e contexto adicional
-        response = model.generate_content(f'{pergunta}. {dados}. Considere que o salário mínimo é R$1412 e dê respostas concisas.', generation_config = genai.GenerationConfig(
-        temperature=0.1,
-    )
-)
-        print(response.text)
-
-# Conectando a função ao botão
-botao_enviar.on_click(on_botao_enviar_clicked)
-
-# Exibindo os widgets
-display(pergunta_texto, botao_enviar, saida_resposta)
-#
